color picker

+++ TwigCodesUI/src/main/java/com/twigcodes/ui/ColorPickerView.kt
@@ -1,7 +1,10 @@
 package com.twigcodes.ui
 
 import android.content.Context
-import android.graphics.*
+import android.graphics.Canvas
+import android.graphics.Color
+import android.graphics.Paint
+import android.graphics.SweepGradient
 import android.util.AttributeSet
 import android.util.Log
 import android.view.MotionEvent
@@ -12,6 +15,33 @@ import com.jakewharton.rxbinding3.view.touches
 import com.twigcodes.ui.util.RxUtil
 import kotlin.math.*
 
+/**
+ *      |
+ *   3  |  4
+ * ----------->
+ *   2  |  1
+ *      |
+ *      ∨
+ * 首先看一下上面的这个坐标系, [SweepGradient]会将color数组按照1-2-3-4象限顺时针分布.
+ * 再来看一下touch时的坐标(event.x, event.y)与(centerX, centerY)差值的正负值, 也是符合这个坐标轴的,
+ * 1-2-3-4象限分别为(+,+), (-,+), (-,-), (+,-).
+ *
+ * 接下来我们需要解决的就是如何将这个差值与color数组的索引对应上.
+ * 单就(x,y)的值来说, 从某种程度上来说它是没有意义的, 因为它只体现了距离, 我们需要的是"角度".
+ * 这里需要简单介绍一下"极坐标系", 它表示点的位置并不是(x, y), 而是(r, θ), r为极径, 可以理解为点P(x,y)到(0,0)的距离,
+ * θ为极角, 可以理解为OX与OP的夹角.
+ * [atan2]方法用于获取θ的值, 这个值其实就是arctan(y/x), 取值范围为(-π,π].
+ * 1->2象限的取值为0->π, 3->4象限的取值为-π->0
+ * 我们需要将这个取值"加工"一下:
+ * (1) 除以2π, 1->2象限的取值变为0->0.5, 3->4象限的取值变为-0.5->0
+ * (2) 如果为负数则加1, 1->2象限的取值变为0->0.5, 3->4象限的取值变为0.5->1
+ * 这样1-2-3-4象限的取值变为0->1, 这与"color数组按照1-2-3-4象限顺时针分布"就对应上了, 于是我们就可以知道touch了
+ * color数组的哪个位置了.
+ *
+ * 接下来我们需要解决的就是颜色的"生成".
+ * 上面获得的值可以理解为是个百分比, 举个例子, 该值为0.4, color数组size为7, 那么0.4*(7-1)=2.4, 也就是当前的选中的颜色
+ * 介于color[2]和color[3]之间, 这个颜色有color[2]60%的基因, 有color[3]40%的基因.
+ */
 class ColorPickerView @JvmOverloads constructor(context: Context, attrs: AttributeSet? = null, defStyleAttr: Int = 0, defStyleRes: Int = 0) : View(context, attrs, defStyleAttr, defStyleRes) {
     companion object {
         private const val PALETTE_STROKE_WIDTH = 32f
@@ -76,8 +106,7 @@ class ColorPickerView @JvmOverloads constructor(context: Context, attrs: Attribu
                                     invalidate()
                                 }
                             } else {
-                                val angle = atan2(y.toDouble(), x.toDouble()).toFloat()
-                                // need to turn angle [-PI ... PI] into unit [0....1]
+                                val angle = atan2(y, x)
                                 // need to turn angle [-PI ... PI] into unit [0....1]
                                 var unit: Float = angle / (2 * Math.PI.toFloat())
                                 if (unit < 0) {
@@ -120,20 +149,6 @@ class ColorPickerView @JvmOverloads constructor(context: Context, attrs: Attribu
         }
     }
 
-//    private fun floatToByte(x: Float): Int {
-//        return Math.round(x)
-//    }
-//
-//    private fun pinToByte(n: Int): Int {
-//        var n = n
-//        if (n < 0) {
-//            n = 0
-//        } else if (n > 255) {
-//            n = 255
-//        }
-//        return n
-//    }
-
     private fun ave(s: Int, d: Int, p: Float): Int {
         return s + (p * (d - s)).roundToInt()
     }
@@ -158,6 +173,20 @@ class ColorPickerView @JvmOverloads constructor(context: Context, attrs: Attribu
         val b = ave(Color.blue(c0), Color.blue(c1), p)
         return Color.argb(a, r, g, b)
     }
+
+//    private fun floatToByte(x: Float): Int {
+//        return Math.round(x)
+//    }
+//
+//    private fun pinToByte(n: Int): Int {
+//        var n = n
+//        if (n < 0) {
+//            n = 0
+//        } else if (n > 255) {
+//            n = 255
+//        }
+//        return n
+//    }
 
 //    private fun rotateColor(color: Int, rad: Float): Int {
 //        val deg = rad * 180 / 3.1415927f
